Module renaming for consistency hydra_graph -> graph_init
